packages/common-python/pyproject.toml
[project]
name = "hephae-common"
version = "0.1.0"
description = "Shared Python code for hephae-forge monorepo"
requires-python = ">=3.11"
dependencies = [
    "firebase-admin>=6.0",
    "google-cloud-bigquery>=3.0",
    "google-cloud-storage>=2.0",
    "pydantic>=2.0",
    "fastapi>=0.100",
    "google-genai>=1.0",
    "google-adk>=0.1",
    "resend>=2.5.0",
    "playwright>=1.48.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hephae_common"]
